backend/requirements.txt
pip>=25.0
setuptools<70
wheel

fastapi==0.115.6
starlette==0.41.3
uvicorn==0.34.0

pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1

openai==1.59.7
httpx==0.27.2
requests==2.32.3
orjson==3.10.14
anyio==4.6.0
sniffio==1.3.1

psycopg[binary]==3.3.1
psycopg_pool>=3.0
numpy==1.26.4
pgvector==0.4.0

openai==2.15.0